backend/requirements.txt

Flask==3.0.0
Flask-CORS==4.0.0
gunicorn==21.2.0

azure-cosmos==4.5.1
azure-storage-blob==12.19.0
azure-functions==1.18.0
azure-identity==1.15.0

openai==1.10.0

numpy==1.25.2
pandas==2.1.4
requests==2.31.0
beautifulsoup4==4.12.2


python-dotenv==1.0.0
python-dateutil==2.8.2
Werkzeug==3.0.1

logging==0.4.9.6

PyPDF2==3.0.1
python-docx==1.1.0
openpyxl==3.1.2

pytest==7.4.3
pytest-asyncio==0.21.1
black==23.12.0
flake8==7.0.0